src/main/java/cn/csq/bishe/service/impl/ProductCategoryServiceImpl.java
package cn.csq.bishe.service.impl;

import cn.csq.bishe.dataobject.ProductCategory;
import cn.csq.bishe.repository.ProductCategoryRepository;
import cn.csq.bishe.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目服务实现
 * @ClassName : ProductCategoryServiceImpl
 * @Author : CSQ
 * @Date: 2020-03-11 02:18
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryRepository.getOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
